examples/aws_strands_agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autopep8==2.3.2
aws-requests-auth==0.4.3
beautifulsoup4==4.13.5
bedrock-agentcore==0.1.4
bedrock-agentcore-starter-toolkit==0.1.12
boto3==1.40.34
botocore==1.40.34
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
dill==0.4.0
docstring_parser==0.17.0
dotenv==0.9.9
frozenlist==1.7.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
mcp==1.23.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pathable==0.4.4
pillow==12.1.1
prance==25.4.8.0
prompt_toolkit==3.0.52
propcache==0.3.2
py-openapi-schema-to-json-schema==0.0.3
pycodestyle==2.14.0
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.22
PyYAML==6.0.2
questionary==2.1.1
readabilipy==0.3.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
slack_bolt==1.25.0
slack_sdk==3.36.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
starlette==0.49.1
strands-agents==1.9.0
strands-agents-tools==0.2.8
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
